Рефакторинг модуля utils
