+++ src/utils/utils.js
@@ -15,29 +15,6 @@ export const getRandomArray = (array, start, end = array.length) => {
     .map(() => getRandomArrayItem(array));
 };
 
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
-export const renderElement = (container, component, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(component.getElement());
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(component.getElement());
-      break;
-  }
-};
-
-export const createElement = (template) => {
-  const element = document.createElement(`div`);
-  element.innerHTML = template;
-
-  return element.firstChild;
-};
-
 export const parseTime = (UTCTimestamp) => {
   const date = new Date(UTCTimestamp);
   return `${date.getHours()}:${date.getMinutes()}`;
